Removed L3 and fixed up some of the SiFiveP550 files
